requirements.txt
discord
ffmpeg
yt_dlp
python-dotenv
PyNaCl